modified:   routes.py

+++ routes.py
@@ -127,8 +127,8 @@ def quote():
                 # --- Placeholder for nesting logic in utils/nesting_utils.py ---
                 sheet_width = material.width
                 sheet_length = material.length
-                useful_width = sheet_width - (Setting.query.filter_by(name='useless_margin').first().value * 2)
-                useful_length = sheet_length - (Setting.query.filter_by(name='useless_margin').first().value * 2)
+                useful_width = sheet_width - (Decimal(Setting.query.filter_by(name='useless_margin').first().value * 2))
+                useful_length = sheet_length - (Decimal(Setting.query.filter_by(name='useless_margin').first().value * 2))
                 parts_lenght_1 = useful_length // element_height
                 parts_width_1 = useful_width // element_width
                 parts_lenght_2 = useful_length // element_width
